Adds aot_compilation.py.
